added crud, permission for file model

+++ src/files/files.controller.ts
@@ -10,13 +10,20 @@ import {
   UploadedFile,
   ParseFilePipe,
   MaxFileSizeValidator,
+  FileTypeValidator,
+  UseGuards,
+  Req,
+  ForbiddenException,
 } from '@nestjs/common';
 import { FilesService } from './files.service';
-import { CreateFileDto } from './dto/create-file.dto';
 import { UpdateFileDto } from './dto/update-file.dto';
 import { ApiBody, ApiConsumes, ApiTags } from '@nestjs/swagger';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { fileStorage } from './storage';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { ApiBearerAuth } from '@nestjs/swagger';
+import { AuthReq } from 'src/contracts/authInterfaces';
+import { FORBIDDEN_MESSAGE } from '@nestjs/core/guards';
 
 @Controller('files')
 @ApiTags('files')
@@ -24,8 +31,10 @@ export class FilesController {
   constructor(private readonly filesService: FilesService) {}
 
   @Post()
+  @UseGuards(JwtAuthGuard)
   @UseInterceptors(FileInterceptor('file', { storage: fileStorage }))
   @ApiConsumes('multipart/form-data')
+  @ApiBearerAuth()
   @ApiBody({
     schema: {
       type: 'object',
@@ -37,34 +46,54 @@ export class FilesController {
       },
     },
   })
-  create(
+  async create(
     @UploadedFile(
       new ParseFilePipe({
-        validators: [new MaxFileSizeValidator({ maxSize: 1024 * 1024 * 5 })],
+        validators: [
+          new FileTypeValidator({ fileType: '.(png|jpeg|jpg)' }),
+          new MaxFileSizeValidator({ maxSize: 1024 * 1024 * 5 }),
+        ],
       }),
     )
     file: Express.Multer.File,
+    @Req() req: AuthReq,
   ) {
-    return file;
+    const user = req.user;
+    return await this.filesService.create(user, file);
   }
 
   @Get()
-  findAll() {
-    return this.filesService.findAll();
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  async findAll(@Req() req: AuthReq) {
+    const user = req.user;
+    return await this.filesService.findAll(user);
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.filesService.findOne(+id);
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  findOne(@Param('id') id: string, @Req() req: AuthReq) {
+    const user = req.user;
+    return this.filesService.findByIdAndUser(+id, user);
   }
 
   @Patch(':id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   update(@Param('id') id: string, @Body() updateFileDto: UpdateFileDto) {
     return this.filesService.update(+id, updateFileDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.filesService.remove(+id);
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  async remove(@Param('id') id: string, @Req() req: AuthReq) {
+    const user = req.user;
+    const file = await this.filesService.findOneById(+id);
+    if (!file || file.user.id !== user.id) {
+      throw new ForbiddenException();
+    }
+    return await this.filesService.remove(+id);
   }
 }
